Reconsitution whole emulator system
1. split Block, Packet, Link, Sender into individual file;
2. change the name "package" to "packet";
3. change the name of player's code file to "congestion_control_algorithm.py";
4. change the name "Network" to "Engine";
5. change the function name "push_to_player" to "append_cc_input".
