+++ objects/block.py
@@ -1,0 +1,35 @@
+from utils import get_ms_time
+
+
+class Block(object):
+
+    def __init__(self,
+                 priority=0,
+                 block_id=-1,
+                 bytes_size=200000,
+                 deadline=0.2,
+                 timestamp=None):
+
+        self.priority = priority
+        self.block_id = block_id
+        self.size = bytes_size
+        self.deadline = deadline
+        self.timestamp = timestamp if not timestamp is None else get_ms_time()
+        # emulator params
+        self.queue_ms = -1
+        self.propagation_ms = -1
+        self.transmition_ms = -1
+
+        # log params
+        self.finish_timestamp = -1
+        self.miss_ddl = 0
+
+
+    def get_cost_time(self):
+
+        return self.queue_ms + self.transmition_ms + self.propagation_ms
+
+
+    def __str__(self):
+
+        return str(self.__dict__)
