+++ pcc_emulator.py
@@ -10,140 +10,17 @@ import numpy as np
 currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
 parentdir = os.path.dirname(currentdir)
 sys.path.insert(0, parentdir)
-from common import sender_obs
+
 from utils import (
-    analyze_pcc_emulator, Package, get_emulator_info,
-    check_solution_format,
-    Appication_Layer
+    analyze_pcc_emulator, get_emulator_info
 )
-from player import Solution
+from objects.sender import Sender
+from objects.link import Link
 
+from congestion_control_algorithm import Solution
+from config.constant import *
 
-DELTA_SCALE = 0.9
-
-MAX_CWND = 5000
-MIN_CWND = 4
-
-MAX_RATE = 1000
-MIN_RATE = 40
-
-REWARD_SCALE = 0.001
-
-EVENT_TYPE_SEND = 'S'
-EVENT_TYPE_ACK = 'A'
-
-BYTES_PER_PACKET = 1500
-HEAD_PER_PACKAET = 20
-
-LATENCY_PENALTY = 1.0
-LOSS_PENALTY = 1.0
-
-USE_LATENCY_NOISE = False
-MAX_LATENCY_NOISE = 1.1
-
-USE_CWND = True
-
-
-class Link():
-
-    def __init__(self, trace_list, queue_size):
-        '''
-        :param trace_list: [[time, bandwith, loss_rate, delay] ...]
-        :param queue_size:
-        '''
-        self.trace_list = trace_list
-        if len(trace_list) == 0:
-            self.bw = 20
-            self.lr = .0
-            self.dl = .0
-        else:
-            self.bw = trace_list[0][1] * 10**6 / BYTES_PER_PACKET
-            self.lr = trace_list[0][2]
-            self.dl = trace_list[0][3]
-
-        self.queue_delay = 0.0
-        self.queue_delay_update_time = 0.0
-        self.max_queue_delay = queue_size / self.bw
-
-        self.extra_delay = 0
-        self.queue_size = queue_size
-
-
-    def get_cur_queue_delay(self, event_time):
-        return max(0.0, self.queue_delay - (event_time - self.queue_delay_update_time))
-
-
-    def get_cur_latency(self, event_time):
-        return self.dl + self.get_cur_queue_delay(event_time)
-
-
-    def packet_enters_link(self, event_time):
-        if (random.random() < self.lr):
-            return False
-        self.queue_delay = self.get_cur_queue_delay(event_time)
-        self.queue_delay_update_time = event_time
-        self.extra_delay = self.send_delay(event_time) # 1.0 / self.bw
-        # print("Extra delay: %f, Current delay: %f, Max delay: %f" % (extra_delay, self.queue_delay, self.max_queue_delay))
-        if self.extra_delay + self.queue_delay > self.max_queue_delay:
-            # print("\tDrop!")
-            return False
-        self.queue_delay += self.extra_delay
-        # print("\tNew delay = %f" % self.queue_delay)
-        return True
-
-
-    def update_trace(self, event_time):
-
-        while len(self.trace_list) > 0 and \
-                event_time > self.trace_list[0][0]:
-            self.trace_list.pop(0)
-
-
-    def send_delay(self, event_time):
-
-        rest_block_size = 1
-        transmition_ms = 0
-        # different bw
-        for i in range(len(self.trace_list)):
-            if rest_block_size <= 0:
-                break
-            if event_time + transmition_ms > self.trace_list[i][0]:
-                continue
-
-            used_time = rest_block_size / self.bw
-            tmp = self.trace_list[i][0] - (event_time + transmition_ms)
-            if used_time > tmp:
-                used_time = tmp
-                rest_block_size -= used_time * self.bw
-                self.bw = self.trace_list[i][1] * 10 ** 6 / BYTES_PER_PACKET
-            else:
-                rest_block_size = 0
-            transmition_ms += used_time
-
-        if rest_block_size > 0:
-            transmition_ms += rest_block_size / self.bw
-            self.update_trace(event_time+transmition_ms)
-
-        self.max_queue_delay = self.queue_size / self.bw
-
-        return transmition_ms
-
-
-    def print_debug(self):
-        print("Link:")
-        print("Bandwidth: %f" % self.bw)
-        print("Delay: %f" % self.dl)
-        print("Queue Delay: %f" % self.queue_delay)
-        print("Max Queue Delay: %f" % self.max_queue_delay)
-        print("One Packet Queue Delay: %f" % (1.0 / self.bw))
-
-
-    def reset(self):
-        self.queue_delay = 0.0
-        self.queue_delay_update_time = 0.0
-
-
-class Network():
+class Engine():
 
     def __init__(self, senders, links):
         self.q = []
@@ -153,16 +30,16 @@ class Network():
         self.queue_initial_packets()
 
         self.fir_log = True
-        self.log_package_file = "output/pcc_emulator_package.log"
+        self.log_packet_file = "output/pcc_emulator_packet.log"
 
 
     def queue_initial_packets(self):
         for sender in self.senders:
             sender.register_network(self)
             sender.reset_obs()
-            package = sender.new_package(1.0 / sender.rate)
-            if package:
-                heapq.heappush(self.q, (1.0 / sender.rate, sender, package))
+            packet = sender.new_packet(1.0 / sender.rate)
+            if packet:
+                heapq.heappush(self.q, (1.0 / sender.rate, sender, packet))
 
 
     def reset(self):
@@ -188,11 +65,11 @@ class Network():
                 print("There is no packet from application~")
                 exit(0)
 
-            event_time, sender, package = queue_item
-            self.log_package(event_time, package)
-            self.push_to_player(*queue_item)
+            event_time, sender, packet = queue_item
+            self.log_packet(event_time, packet)
+            self.append_cc_input(*queue_item)
 
-            event_type, next_hop, cur_latency, dropped, life, package_id = package.parse()
+            event_type, next_hop, cur_latency, dropped, life, packet_id = packet.parse()
             # print("Got event %s, to link %d, latency %f at time %f" % (event_type, next_hop, cur_latency, event_time))
             self.cur_time = event_time
             new_event_time = event_time
@@ -209,7 +86,7 @@ class Network():
                         sender.on_packet_lost()
                         # print("Packet lost at time %f" % self.cur_time)
                     else:
-                        sender.on_packet_acked(cur_latency, package)
+                        sender.on_packet_acked(cur_latency, packet)
                         # print("Packet acked at time %f" % self.cur_time)
                 # ack back to source
                 else:
@@ -226,9 +103,9 @@ class Network():
                     if sender.can_send_packet():
                         sender.on_packet_sent()
                         push_new_event = True
-                    _package = sender.new_package(self.cur_time + (1.0 / sender.rate))
-                    if _package:
-                        heapq.heappush(self.q, (self.cur_time + (1.0 / sender.rate), sender, _package))
+                    _packet = sender.new_packet(self.cur_time + (1.0 / sender.rate))
+                    if _packet:
+                        heapq.heappush(self.q, (self.cur_time + (1.0 / sender.rate), sender, _packet))
 
                 else:
                     push_new_event = True
@@ -246,11 +123,11 @@ class Network():
                 life += link_latency + sender.path[next_hop].extra_delay
 
             if push_new_event:
-                package.next_hop = new_next_hop
-                package.package_type = new_event_type
-                package.queue_delay = new_latency
-                package.drop = new_dropped
-                heapq.heappush(self.q, (new_event_time, sender, package))
+                packet.next_hop = new_next_hop
+                packet.packet_type = new_event_type
+                packet.queue_delay = new_latency
+                packet.drop = new_dropped
+                heapq.heappush(self.q, (new_event_time, sender, packet))
 
         sender_mi = self.senders[0].get_run_data()
         throughput = sender_mi.get("recv rate")
@@ -280,35 +157,35 @@ class Network():
         return reward * REWARD_SCALE
 
 
-    def log_package(self, event_time, package):
+    def log_packet(self, event_time, packet):
         '''
-        package is tuple of (event_time, sender, event_type, next_hop, cur_latency, dropped, package_id, life)
-        :param package: tuple
-        :return: package
+        packet is tuple of (event_time, sender, event_type, next_hop, cur_latency, dropped, packet_id, life)
+        :param packet: tuple
+        :return: Packet
         '''
 
         if self.fir_log:
             self.fir_log = False
-            with open(self.log_package_file, "w") as f:
+            with open(self.log_packet_file, "w") as f:
                 pass
 
         log_data = { "Time" : event_time }
-        log_data.update(package.trans2dict())
+        log_data.update(packet.trans2dict())
 
-        with open(self.log_package_file, "a") as f:
+        with open(self.log_packet_file, "a") as f:
             f.write(str(log_data)+"\n")
 
-        return package
+        return packet
 
 
-    def push_to_player(self, event_time, sender, package, event_type="package"):
+    def append_cc_input(self, event_time, sender, packet, event_type="packet"):
 
-        if event_type == "package":
+        if event_type == "packet":
             data = {
                 "event_time" : event_time,
-                "link_rate" : -1 if package.next_hop == 0 else sender.path[package.next_hop-1].bw,
+                "link_rate" : -1 if packet.next_hop == 0 else sender.path[packet.next_hop-1].bw,
                 "send_rate" : sender.rate,
-                "package" : package.trans2dict()
+                "packet" : packet.trans2dict()
             }
         elif event_type == "system":
             data = {
@@ -316,182 +193,6 @@ class Network():
             }
 
         sender.solution.input_list.append(data)
-
-
-class Sender():
-
-    def __init__(self, path, dest, features, history_len=10, solution=None):
-        self.id = Sender._get_next_id()
-        self.sent = 0
-        self.acked = 0
-        self.lost = 0
-        self.bytes_in_flight = 0
-        self.min_latency = None
-        self.rtt_samples = []
-        self.sample_time = []
-        self.net = None
-        self.path = path
-        self.dest = dest
-        self.history_len = history_len
-        self.features = features
-        self.history = sender_obs.SenderHistory(self.history_len,
-                                                self.features, self.id)
-
-        self.solution = solution
-        ret = check_solution_format(self.solution.make_decision())
-        self.rate = ret["send_rate"]
-        self.cwnd = ret["cwnd"]
-        self.starting_rate = self.rate
-
-        self.application = None
-
-
-    _next_id = 1
-
-    @classmethod
-    def _get_next_id(cls):
-        result = Sender._next_id
-        Sender._next_id += 1
-        return result
-
-
-    def init_application(self, block_file):
-        self.application = Appication_Layer(block_file, BYTES_PER_PACKET)
-
-
-    def new_package(self, cur_time):
-        package = self.application.get_next_package(cur_time)
-        if package:
-            package.send_delay = 1 / self.rate
-        # print(package)
-        return package
-
-
-    def apply_rate_delta(self, delta):
-        delta *= DELTA_SCALE
-        # print("Applying delta %f" % delta)
-        if delta >= 0.0:
-            self.set_rate(self.rate * (1.0 + delta))
-        else:
-            self.set_rate(self.rate / (1.0 - delta))
-
-
-    def apply_cwnd_delta(self, delta):
-        delta *= DELTA_SCALE
-        # print("Applying delta %f" % delta)
-        if delta >= 0.0:
-            self.set_cwnd(self.cwnd * (1.0 + delta))
-        else:
-            self.set_cwnd(self.cwnd / (1.0 - delta))
-
-
-    def can_send_packet(self):
-        if USE_CWND:
-            ret = self.solution.make_decision()
-            self.rate, self.cwnd = ret["send_rate"], ret["cwnd"]
-            return int(self.bytes_in_flight) / BYTES_PER_PACKET < self.cwnd
-        else:
-            return True
-
-
-    def register_network(self, net):
-        self.net = net
-
-
-    def on_packet_sent(self):
-        self.sent += 1
-        self.bytes_in_flight += BYTES_PER_PACKET
-
-
-    def on_packet_acked(self, rtt, package):
-        self.acked += 1
-        self.rtt_samples.append(rtt)
-        if (self.min_latency is None) or (rtt < self.min_latency):
-            self.min_latency = rtt
-        self.bytes_in_flight -= BYTES_PER_PACKET
-        self.application.get_ack_package(package)
-
-
-    def on_packet_lost(self):
-        self.lost += 1
-        self.bytes_in_flight -= BYTES_PER_PACKET
-
-
-    def set_rate(self, new_rate):
-        self.rate = new_rate
-        # print("Attempt to set new rate to %f (min %f, max %f)" % (new_rate, MIN_RATE, MAX_RATE))
-        if self.rate > MAX_RATE:
-            self.rate = MAX_RATE
-        if self.rate < MIN_RATE:
-            self.rate = MIN_RATE
-
-
-    def set_cwnd(self, new_cwnd):
-        self.cwnd = int(new_cwnd)
-        # print("Attempt to set new rate to %f (min %f, max %f)" % (new_rate, MIN_RATE, MAX_RATE))
-        if self.cwnd > MAX_CWND:
-            self.cwnd = MAX_CWND
-        if self.cwnd < MIN_CWND:
-            self.cwnd = MIN_CWND
-
-
-    def record_run(self):
-        smi = self.get_run_data()
-        self.history.step(smi)
-
-
-    def get_obs(self):
-        return self.history.as_array()
-
-
-    def get_run_data(self):
-        obs_end_time = self.net.get_cur_time()
-
-        # obs_dur = obs_end_time - self.obs_start_time
-        # print("Got %d acks in %f seconds" % (self.acked, obs_dur))
-        # print("Sent %d packets in %f seconds" % (self.sent, obs_dur))
-        # print("self.rate = %f" % self.rate)
-        # print(self.rtt_samples)
-        return sender_obs.SenderMonitorInterval(
-            self.id,
-            bytes_sent=self.sent * BYTES_PER_PACKET,
-            bytes_acked=self.acked * BYTES_PER_PACKET,
-            bytes_lost=self.lost * BYTES_PER_PACKET,
-            send_start=self.obs_start_time,
-            send_end=obs_end_time,
-            recv_start=self.obs_start_time,
-            recv_end=obs_end_time,
-            rtt_samples=self.rtt_samples,
-            packet_size=BYTES_PER_PACKET
-        )
-
-
-    def reset_obs(self):
-        self.sent = 0
-        self.acked = 0
-        self.lost = 0
-        self.rtt_samples = []
-        self.obs_start_time = self.net.get_cur_time()
-
-
-    def print_debug(self):
-        print("Sender:")
-        print("Obs: %s" % str(self.get_obs()))
-        print("Rate: %f" % self.rate)
-        print("Sent: %d" % self.sent)
-        print("Acked: %d" % self.acked)
-        print("Lost: %d" % self.lost)
-        print("Min Latency: %s" % str(self.min_latency))
-
-
-    def reset(self):
-        # print("Resetting sender!")
-        self.rate = self.starting_rate
-        self.bytes_in_flight = 0
-        self.min_latency = None
-        self.reset_obs()
-        self.history = sender_obs.SenderHistory(self.history_len,
-                                                self.features, self.id)
 
 
 class PccEmulator(object):
@@ -515,7 +216,7 @@ class PccEmulator(object):
         self.links = None
         self.senders = None
         self.create_new_links_and_senders()
-        self.net = Network(self.senders, self.links)
+        self.net = Engine(self.senders, self.links)
 
         # for player
         self.solution = Solution()
@@ -544,7 +245,7 @@ class PccEmulator(object):
         # queue = 1 + int(np.exp(random.uniform(*self.queue_range)))
         # print("queue size : %d" % queue)
         # bw = self.trace_list[0][1]
-        bw    = 705 # true bw is bw*BYTES_PER_PACKAGE
+        bw    = 705 # true bw is bw*BYTES_PER_PACKET
         lat   = 0.03
         queue = 5
         loss  = 0.00
@@ -593,7 +294,7 @@ class PccEmulator(object):
         self.steps_taken = 0
         self.net.reset()
         self.create_new_links_and_senders()
-        self.net = Network(self.senders, self.links)
+        self.net = Engine(self.senders, self.links)
         self.episodes_run += 1
         if self.episodes_run > 0 and self.episodes_run % 100 == 0:
             self.dump_events_to_file("pcc_env_log_run_%d.json" % self.episodes_run)
@@ -630,7 +331,7 @@ if __name__ == '__main__':
     block_file = "config/block.txt"
     trace_file = "config/trace.txt"
     log_file = "output/pcc_emulator.log"
-    log_package_file = "output/pcc_emulator_package.log"
+    log_packet_file = "output/pcc_emulator_packet.log"
 
     emulator = PccEmulator(
         block_file=block_file,
@@ -642,4 +343,4 @@ if __name__ == '__main__':
     emulator.print_debug()
     print(emulator.senders[0].rtt_samples)
     print(emulator.senders[0].application.ack_blocks)
-    analyze_pcc_emulator(log_package_file, trace_file)
+    analyze_pcc_emulator(log_packet_file, trace_file)
