+++ utils.py
@@ -17,274 +17,6 @@ def get_ms_time(rate=1000):
     return time.time()*rate
 
 
-class Block(object):
-
-    def __init__(self,
-                 priority=0,
-                 block_id=-1,
-                 bytes_size=200000,
-                 deadline=0.2,
-                 timestamp=None):
-
-        self.priority = priority
-        self.block_id = block_id
-        self.size = bytes_size
-        self.deadline = deadline
-        self.timestamp = timestamp if not timestamp is None else get_ms_time()
-        # emulator params
-        self.queue_ms = -1
-        self.propagation_ms = -1
-        self.transmition_ms = -1
-
-        # log params
-        self.finish_timestamp = -1
-        self.miss_ddl = 0
-
-
-    def get_cost_time(self):
-
-        return self.queue_ms + self.transmition_ms + self.propagation_ms
-
-
-    def __str__(self):
-
-        return str(self.__dict__)
-
-
-class Package(object):
-    _package_id = 1
-
-    def __init__(self,
-                 create_time,
-                 next_hop,
-                 block_id,
-                 offset,
-                 payload,
-                 package_id=None,
-                 package_size=1500,
-                 deadline=0.2,
-                 package_type="S",
-                 drop = False,
-                 send_delay=.0,
-                 queue_delay=.0
-                 ):
-        self.package_type = package_type
-        self.create_time = create_time
-        self.next_hop = next_hop
-        self.block_id = block_id
-        self.offset = offset
-        self.package_id = package_id
-        self.payload=payload
-        self.package_size=package_size
-        self.deadline = deadline
-        self.drop = drop
-
-        self.send_delay = send_delay
-        self.queue_delay = queue_delay
-        self.propagation_delay = 0.0002
-
-        if package_id is None:
-            self.package_id = Package._get_next_package()
-
-
-    @classmethod
-    def _get_next_package(cls):
-        result = cls._package_id
-        cls._package_id += 1
-        return result
-
-
-    def parse(self):
-
-        return [self.package_type,
-                self.next_hop,
-                self.queue_delay,
-                self.drop,
-                self.queue_delay,
-                self.package_id]
-
-
-    def trans2dict(self):
-        print_data = {
-            "Type": self.package_type,
-            "Position": self.next_hop,
-            "Send_delay": self.send_delay,
-            "Queue_delay": self.queue_delay,
-            "Propagation_delay": self.propagation_delay,
-            "Drop": 1 if self.drop else 0,
-            "Package_id": self.package_id,
-            "Block_id": self.block_id,
-            "Create_time": self.create_time,
-            "Deadline": self.deadline,
-            "Offset" : self.offset,
-            "Payload" : self.payload,
-            "Package_size" : self.package_size
-        }
-        return print_data
-
-
-    def __lt__(self, other):
-        return self.create_time < other.create_time
-
-
-    def __str__(self):
-        print_data = self.trans2dict()
-        return str(print_data)
-
-
-class Appication_Layer(object):
-
-
-    def __init__(self,
-                 block_file,
-                 bytes_per_packet=1500):
-        self.block_file = block_file
-        self.block_queue = []
-        self.bytes_per_packet = bytes_per_packet
-
-        self.block_nums = None
-        self.split_nums = 0
-        self.init_time = .0
-        self.pass_time = .0
-        self.fir_log = True
-
-        self.now_block = None
-        self.now_block_offset = 0
-        self.head_per_packet = 20
-
-        self.create_block_by_file()
-        self.ack_blocks = {}
-
-
-    def create_block_by_file(self, det=1):
-        with open(self.block_file, "r") as f:
-            self.block_nums = int(f.readline())
-
-            pattern_cols = ["type", "size", "ddl"]
-            pattern=[]
-            for line in f.readlines():
-                pattern.append(
-                    { pattern_cols[idx]:item.strip() for idx, item in enumerate(line.split(',')) }
-                )
-
-            peroid = len(pattern)
-            for idx in range(self.block_nums):
-                ch = idx % peroid
-                block = Block(bytes_size=float(pattern[ch]["size"]),
-                              block_id=idx,
-                              deadline=float(pattern[ch]["ddl"]),
-                              timestamp=self.init_time+self.pass_time+idx*det,
-                              priority=pattern[ch]["type"])
-                self.block_queue.append(block)
-
-
-    def select_block(self):
-
-        def is_better(block):
-            return (now_time - block.timestamp) * best_block.deadline > \
-                    (now_time - best_block.timestamp) * block.deadline
-
-
-        now_time = self.init_time + self.pass_time
-        best_block = None
-        ch = -1
-        need_filter = []
-        for idx, item in enumerate(self.block_queue):
-            # if miss ddl in queue, clean and log
-            if now_time > item.timestamp + item.deadline:
-                self.block_queue[idx].miss_ddl = 1
-                self.log_block(self.block_queue[idx])
-                need_filter.append(idx)
-                print(now_time, item.timestamp, item.deadline)
-
-            elif best_block == None or is_better(item) :
-                best_block = item
-                ch = idx
-
-        # filter block with missing ddl
-        for idx in range(len(need_filter)-1, -1, -1):
-            if ch != -1 and ch > idx:
-                self.block_queue.pop(ch)
-                ch = -1
-            self.block_queue.pop(need_filter[idx])
-        if ch != -1:
-            self.block_queue.pop(ch)
-        return best_block
-
-
-    def get_retrans_package(self):
-        for key, val in self.ack_blocks.items():
-            if len(val) == self.split_nums:
-                continue
-            for i in range(self.split_nums):
-                if i not in val:
-                    return i
-        return None
-
-
-    def get_next_package(self, cur_time):
-        self.pass_time = cur_time
-        retrans_package = None
-        if self.now_block is None or self.now_block_offset == self.split_nums:
-            # 1. the retransmisson time is bad, which may cause consistently loss packet
-            # 2. the packet will be retransmission many times for a while
-            retrans_package = self.get_retrans_package()
-            if retrans_package:
-                self.now_block_offset = retrans_package
-            else:
-                self.now_block = self.select_block()
-                if self.now_block is None:
-                    return None
-
-                self.now_block_offset = 0
-                self.split_nums = int(np.ceil(self.now_block.size / self.bytes_per_packet))
-
-        payload = self.bytes_per_packet - self.head_per_packet
-        if self.now_block.size % (self.bytes_per_packet - self.head_per_packet) and \
-                self.now_block_offset == self.split_nums - 1:
-            payload = self.now_block.size % (self.bytes_per_packet - self.head_per_packet)
-
-        package = Package(create_time=cur_time,
-                          next_hop=0,
-                          block_id=self.now_block.block_id,
-                          offset=self.now_block_offset,
-                          package_size=self.bytes_per_packet,
-                          payload=payload
-                          )
-        if retrans_package is None:
-            self.now_block_offset += 1
-        else:
-            self.now_block_offset = self.split_nums
-
-        return package
-
-
-    def get_ack_package(self, package):
-        if package.block_id not in self.ack_blocks:
-            self.ack_blocks[package.block_id] = [package.offset]
-        else:
-            self.ack_blocks[package.block_id].append(package.offset)
-
-    # todo : Adjust log time and content
-    def log_block(self, block):
-
-        if self.fir_log:
-            self.fir_log = False
-            with open("output/block.log", "w") as f:
-                pass
-
-        block.finish_timestamp = self.init_time + self.pass_time
-        if block.get_cost_time() > block.deadline:
-            block.miss_ddl = 1
-
-        with open("output/block.log", "a") as f:
-            f.write(str(block)+'\n')
-
-
-    def analyze_application(self):
-        pass
-
-
 def analyze_pcc_emulator(log_file, trace_file=None, rows=20):
 
     plt_data = []
@@ -294,8 +26,8 @@ def analyze_pcc_emulator(log_file, trace_file=None, rows=20):
             plt_data.append(json.loads(line.replace("'", '"')))
 
     plt_data = filter(lambda x:x["Type"]=='A' and x["Position"] == 2, plt_data)
-    # priority by package id
-    plt_data = sorted(plt_data, key=lambda x:int(x["Package_id"]))
+    # priority by packet id
+    plt_data = sorted(plt_data, key=lambda x:int(x["Packet_id"]))
 
     pic_nums = 3
     data_lantency = []
@@ -316,7 +48,7 @@ def analyze_pcc_emulator(log_file, trace_file=None, rows=20):
     plt.figure(figsize=(20, 5*pic_nums))
     # plot latency distribution
     ax = plt.subplot(pic_nums, 1, 1)
-    ax.set_title("Acked package latency distribution", fontsize=30)
+    ax.set_title("Acked packet latency distribution", fontsize=30)
     ax.set_ylabel("Latency / s", fontsize=20)
     ax.set_xlabel("Time / s", fontsize=20)
     ax.scatter(data_finish_time, data_lantency, label="Latency")
@@ -328,7 +60,7 @@ def analyze_pcc_emulator(log_file, trace_file=None, rows=20):
 
     # plot miss deadline rate block
     ax = plt.subplot(pic_nums, 1, 2)
-    ax.set_title("Acked package lost distribution", fontsize=30)
+    ax.set_title("Acked packet lost distribution", fontsize=30)
     ax.set_ylabel("Latency / s", fontsize=20)
     ax.set_xlabel("Time / s", fontsize=20)
     ax.scatter([data_finish_time[idx] for idx in data_drop],
@@ -340,7 +72,7 @@ def analyze_pcc_emulator(log_file, trace_file=None, rows=20):
 
     # plot latency distribution
     ax = plt.subplot(pic_nums, 1, 3)
-    ax.set_title("Acked package life time distribution", fontsize=30)
+    ax.set_title("Acked packet life time distribution", fontsize=30)
     ax.set_ylabel("Latency / s", fontsize=20)
     ax.set_xlabel("Time / s", fontsize=20)
     ax.set_ylim(-np.min(data_sum_time)*2, np.max(data_sum_time)*2)
@@ -412,5 +144,5 @@ def get_emulator_info(sender_mi):
 
 if __name__ == '__main__':
 
-    log_package_file = "output/pcc_emulator_package.log"
-    analyze_pcc_emulator(log_package_file)
+    log_packet_file = "output/pcc_emulator_packet.log"
+    analyze_pcc_emulator(log_packet_file)
